requirements.txt
pypdf
Chroma
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
matplotlib
seaborn
CTransformers
pypdf2
datasets
Flask
DateTime
mysql-connector-python
mysqlclient
pandas